rename sign server and merge conflicts
