+++ main.go
@@ -13,8 +13,8 @@ import (
 	"erupe-ce/server/discordbot"
 	"erupe-ce/server/entranceserver"
 	"erupe-ce/server/launcherserver"
-	"erupe-ce/server/newsignserver"
 	"erupe-ce/server/signserver"
+	"erupe-ce/server/signv2server"
 
 	"github.com/jmoiron/sqlx"
 	_ "github.com/lib/pq"
@@ -168,17 +168,17 @@ func main() {
 	}
 
 	// New Sign server
-	var newSignServer *newsignserver.Server
-	if config.ErupeConfig.NewSign.Enabled {
-		newSignServer = newsignserver.NewServer(
-			&newsignserver.Config{
+	var newSignServer *signv2server.Server
+	if config.ErupeConfig.SignV2.Enabled {
+		newSignServer = signv2server.NewServer(
+			&signv2server.Config{
 				Logger:      logger.Named("sign"),
 				ErupeConfig: config.ErupeConfig,
 				DB:          db,
 			})
 		err = newSignServer.Start()
 		if err != nil {
-			preventClose(fmt.Sprintf("Failed to start new sign server: %s", err.Error()))
+			preventClose(fmt.Sprintf("Failed to start sign-v2 server: %s", err.Error()))
 		}
 		logger.Info("Started new sign server")
 	}
@@ -246,7 +246,7 @@ func main() {
 		signServer.Shutdown()
 	}
 
-	if config.ErupeConfig.NewSign.Enabled {
+	if config.ErupeConfig.SignV2.Enabled {
 		newSignServer.Shutdown()
 	}
